aud: the whole path carries sixteen bits, and only the sink narrows
aud_out was the shape of the problem. It took a uint16_t centered on
AUD_PWM_CENTER, so every driver had to know the RP2350's PWM depth, and
the ten bits that chip can carry travelled to hosts with far better
converters. It now takes a signed sample at full scale and the platform
decides: the RP2350 scales to its PWM, the Pocket hands all sixteen bits
to the I2S, and the emulator makes a float. Only that last step narrows.

Widening the seam alone would have been cosmetic for two of the three
voices, because the ceiling was not the seam. Every oscillator was
int8_t — the sine table both the PSG and the bell read, and the square,
saw, triangle and noise in each of them, in the C and in the RTL alike.
No amount of width downstream lifts a floor set at ±127.

So the oscillators widened too, and for four of the five waveforms that
is a pure resolution change: square is two values scaled, and the ramps
take eight more bits off the same phase accumulator they already used.
That is checkable rather than assertable, and it was checked — the wide
saw and triangle, shifted back down by eight, equal the old ones at all
256 phase steps, wrap included. The int8_t cast in the triangle wrapped
at phase 192 and the int16_t one wraps in the same place. Only the noise
tap genuinely changes texture, taking sixteen bits of the LFSR instead
of eight.

Three arithmetic faults went with it, all of which cost level even on
the RP2350's own ten bits:

The envelope was nine bits of a Q24 accumulator. A long release moved
the gain in 256 visible steps, which is the zipper you could hear on a
slow fade; it takes thirteen now.

The mix truncated twice, once after the envelope and once after the pan,
and then shifted the gap back in as zeros. It accumulates unshifted and
rounds once. A floor is not noise — it is a downward bias every sounding
channel adds to, and eight of them reached -26 dBFS of DC that appeared
and vanished with the notes.

The OPL was thrown from sixteen bits to ten at the source, in opl.c
before any mixer saw it. It multiplies by four now instead of shifting
down by four: same four-times-hot clipping the machine has always had,
without discarding six bits to get there. That exposed a real
disagreement — the fabric was running the OPL at half the C's level,
because SAMPLE_SHIFT was derived from a comment claiming nine voices
reach 2^21 when channels.sv saturates the sum to sixteen bits first.
Three is the shift that matches, and the two now clip on the same note.

And a bug that was never about range: emu8950's mix_output_raw summed
fifteen int16 channels into an int32 and returned it as int16 with no
clamp, on the live OPL_calc_buffer path. A loud chord did not clip, it
inverted.

The tests needed more care than the code. test_aud and test_opl were
subtracting 512 from a value that no longer has 512 as its center, and
their thresholds were loose enough that they kept passing while
measuring the wrong thing; Verilator hands ports over unsigned, so the
first fix read negative samples as 65201 and still passed. Signed and
tightened, they measure a PSG peak of 16128 where one full centred
channel belongs, against a port whose entire old range was 511. test_pi2s
was feeding ten bits scaled up six, which could never have caught a
stage that dropped the bottom six.

Worth knowing for anyone doing this again: cmake --build build does NOT
build tests/fpga. That tree is its own cmake project at build/fpga, and
a full green suite run against it while it is stale looks exactly like a
full green suite run.

+++ tests/fpga/psg_shim.c
@@ -25,10 +25,10 @@ volatile uint8_t xram_queue_head;
 volatile uint8_t xram_queue_tail;
 volatile uint8_t xram_queue[256][2];
 
-int8_t aud_sine_table[256];
+int16_t aud_sine_table[256];
 
 static void (*handler)(void);
-static uint16_t out_l, out_r;
+static int16_t out_l, out_r;
 
 void aud_setup(void (*fn)(void), uint32_t rate)
 {
@@ -36,7 +36,7 @@ void aud_setup(void (*fn)(void), uint32_t rate)
     handler = fn;
 }
 
-void aud_out(uint16_t left, uint16_t right)
+void aud_out(int16_t left, int16_t right)
 {
     out_l = left;
     out_r = right;
@@ -50,7 +50,7 @@ void shim_init(void)
         aud_sine_table[i] = AUD_SINE_TABLE[i];
 }
 
-void shim_sample(uint16_t *l, uint16_t *r)
+void shim_sample(int16_t *l, int16_t *r)
 {
     handler();
     *l = out_l;
